+++ experimental/dds/tree2/api-report/tree2.api.md
@@ -23,6 +23,23 @@ import type { Static } from '@sinclair/typebox';
 import type { TSchema } from '@sinclair/typebox';
 
 // @alpha
+export function adaptEnum<TScope extends string, const TEnum extends Record<string, string>>(factory: SchemaFactory<TScope>, members: TEnum): (<TValue extends TEnum[keyof TEnum]>(value: TValue) => {
+    new (data?: EmptyObject | undefined): {
+        readonly value: TValue;
+    };
+    readonly identifier: `${TScope}.${TValue}`;
+    readonly kind: NodeKind.Object;
+    readonly info: unknown;
+}) & { readonly [Property in keyof TEnum]: {
+        new (data?: EmptyObject | undefined): {
+            readonly value: TEnum[Property];
+        };
+        readonly identifier: `${TScope}.${TEnum[Property]}`;
+        readonly kind: NodeKind.Object;
+        readonly info: unknown;
+    }; };
+
+// @alpha
 export interface Adapters {
     // (undocumented)
     readonly tree?: readonly TreeAdapter[];
@@ -411,6 +428,23 @@ export type EmptyObject = {};
 
 // @alpha
 export function encodeTreeSchema(schema: TreeStoredSchema): JsonCompatible;
+
+// @alpha
+export function enumFromStrings<TScope extends string, const Members extends string>(factory: SchemaFactory<TScope>, members: Members[]): (<TValue extends Members>(value: TValue) => {
+    new (data?: EmptyObject | undefined): {
+        readonly value: TValue;
+    };
+    readonly identifier: `${TScope}.${TValue}`;
+    readonly kind: NodeKind.Object;
+    readonly info: unknown;
+}) & Record<Members, {
+    new (data?: EmptyObject | undefined): {
+        readonly value: Members;
+    };
+    readonly identifier: `${TScope}.${Members}`;
+    readonly kind: NodeKind.Object;
+    readonly info: unknown;
+}>;
 
 // @beta
 export type Events<E> = {
@@ -1603,6 +1637,16 @@ export interface SharedTreeOptions extends Partial<ICodecOptions> {
 export function singleTextCursor(root: JsonableTree): ITreeCursorSynchronous;
 
 // @alpha
+export function singletonSchema<TScope extends string, TName extends string>(factory: SchemaFactory<TScope>, name: TName): {
+    new (data?: EmptyObject): {
+        readonly value: TName;
+    };
+    readonly identifier: `${TScope}.${TName}`;
+    readonly kind: NodeKind.Object;
+    readonly info: unknown;
+};
+
+// @alpha
 export type StableNodeKey = Brand<StableId, "Stable Node Key">;
 
 // @alpha
@@ -1922,6 +1966,9 @@ TFields extends {
     -readonly [key in keyof TFields]: InsertableFlexField<TFields[key]>;
 } : EmptyObject
 ][_InlineTrick];
+
+// @alpha
+export function typedObjectValues<TKey extends string, TValues>(object: Record<TKey, TValues>): TValues[];
 
 // @alpha
 export const typeNameSymbol: unique symbol;
