+++ experimental/dds/tree2/src/class-tree/index.ts
@@ -23,6 +23,13 @@ export { SchemaFactory } from "./schemaFactory";
 export { nodeApi as Tree, TreeApi, TreeNodeEvents } from "./treeApi";
 export { toFlexConfig } from "./toFlexSchema";
 
+export {
+	adaptEnum,
+	enumFromStrings,
+	singletonSchema,
+	typedObjectValues,
+} from "./schemaCreationUtilities";
+
 // Below here are things that are used by the above, but not part of the desired API surface.
 import * as InternalClassTreeTypes from "./internal";
 export { InternalClassTreeTypes };
