+++ experimental/dds/tree2/src/test/tsconfig.json
@@ -4,11 +4,12 @@
 		"lib": ["ES2019", "ES2018.Promise", "DOM", "DOM.Iterable"],
 		"noImplicitAny": true,
 		"noUnusedLocals": false,
+		"noImplicitOverride": true,
 		"rootDir": "./",
 		"outDir": "../../dist/test",
 		"types": ["node", "mocha"],
 	},
-	"include": ["./**/*"],
+	"include": ["./**/*", "../class-tree/schemaCreationUtilities.ts"],
 	"references": [
 		{
 			"path": "../..",
